Improve behaviour for invalid enum keys in Dictionary fields
Previously, invalid enum keys would often crash the program when trying
to add the same key twice.

However, it may not be ideal for us to even add invalid keys to a
Dictionary; so I've provided an option that allows the user to skip
these fields from being added entirely.

In the case of Ignore, it will still set it to the first enum value as
it did before - but it will no longer crash when trying to add the same
key multiple times (in the case where multiple invalid keys were found,
or a valid key at the first enumerator also exists)